+++ export_readiness/tests/factories.py
@@ -64,32 +64,3 @@ class GetFinancePageFactory(wagtail_factories.PageFactory):
     slug_en_gb = factory.Sequence(lambda n: '123-555-{0}'.format(n))
     title_en_gb = factory.Sequence(lambda n: '123-555-{0}'.format(n))
     parent = None
-
-
-class HighPotentialOfferFormPageFactory(wagtail_factories.PageFactory):
-    class Meta:
-        model = models.HighPotentialOfferFormPage
-
-    comment_help_text = factory.fuzzy.FuzzyText(length=200)
-    comment_label = factory.fuzzy.FuzzyText(length=200)
-    company_name_help_text = factory.fuzzy.FuzzyText(length=200)
-    company_name_label = factory.fuzzy.FuzzyText(length=200)
-    company_size_help_text = factory.fuzzy.FuzzyText(length=200)
-    company_size_label = factory.fuzzy.FuzzyText(length=200)
-    country_help_text = factory.fuzzy.FuzzyText(length=200)
-    country_label = factory.fuzzy.FuzzyText(length=200)
-    email_address_help_text = factory.fuzzy.FuzzyText(length=200)
-    email_address_label = factory.fuzzy.FuzzyText(length=200)
-    full_name_help_text = factory.fuzzy.FuzzyText(length=200)
-    full_name_label = factory.fuzzy.FuzzyText(length=200)
-    opportunities_help_text = factory.fuzzy.FuzzyText(length=200)
-    opportunities_label = factory.fuzzy.FuzzyText(length=200)
-    phone_number_help_text = factory.fuzzy.FuzzyText(length=200)
-    phone_number_label = factory.fuzzy.FuzzyText(length=200)
-    role_in_company_help_text = factory.fuzzy.FuzzyText(length=200)
-    role_in_company_label = factory.fuzzy.FuzzyText(length=200)
-    terms_agreed_help_text = factory.fuzzy.FuzzyText(length=200)
-    terms_agreed_label = factory.fuzzy.FuzzyText(length=200)
-    website_url_help_text = factory.fuzzy.FuzzyText(length=200)
-    website_url_label = factory.fuzzy.FuzzyText(length=200)
-    parent = None
