+++ export_readiness/tests/test_models.py
@@ -13,7 +13,6 @@ def test_app_models():
         models.GetFinancePage,
         models.PerformanceDashboardPage,
         models.PerformanceDashboardNotesPage,
-        models.HighPotentialOfferFormPage,
     ]
