+++ export_readiness/translation.py
@@ -32,8 +32,3 @@ class PerformanceDashboardPageTranslationOptions(BaseTranslationOptions):
 @register(models.PerformanceDashboardNotesPage)
 class PerformanceDashboardNotesPageTranslationOptions(BaseTranslationOptions):
     fields = []
-
-
-@register(models.HighPotentialOfferFormPage)
-class HighPotentialOfferFormPageTranslationOptions(BaseTranslationOptions):
-    fields = []
