+++ invest/translation.py
@@ -159,3 +159,8 @@ class SetupGuideLandingPageTranslation(BaseTranslationOptions):
         'sub_heading',
         'lead_in',
     )
+
+
+@register(models.HighPotentialOfferFormPage)
+class HighPotentialOfferFormPageTranslationOptions(BaseTranslationOptions):
+    fields = []
